clean up View initialize order and ElementView doc setting
The intention of the parameters to View constructor is that

    var view = maria.View(model, controller);

is intended to be *exactly* shorthand for

    var view = maria.View();
    view.setModel(model);
    view.setController(controller);

Similarly for the ElementView constructor

   var elementView = maria.ElementView(model, controller, document);

is intended to be *exactly* shorthand for

   var elementView = maria.ElementView();
   elementView.setModel(model);
   elementView.setController(controller);
   elementView.setDocument(document);

Unfortunately the implementation of the maria.View and
maria.ElementView constructors had the call to this.initialize
after the calls to this.setModel, this.setController, and
setting this._doc which makes the program semantics more complex:
a constructor call with parameters is not the same as an empty
constructor call with subsequent setter calls.

This commit makes the constructor call with parameters exactly
equivalent to the shorthand versions shown above. It also means
that when the intialize method is called during construction
that the model, controller, and document have not yet been set
on the model object.

This also matches Squeak more closely.